Add files via upload
This program asks the user to enter an item's wholesale cost and its markup percent-age. It should then display the item's retail price. 
For example:
• If an item's wholesale cost is 5.00 and its markup percentage is 100 percent, then the item's retail price is 10.00.
• If an item's wholesale cost is 5.00 and its markup percentage is 50 percent, then the item's retail price is 7.50.
The program should have a method named calculateRetail that receives the wholesale cost and the markup percentage as arguments, and returns the retail price of the item.